Extract the HTTP status code generating code into a provider.
The JSON-RPC 2.0 spec does not contain information about HTTP status codes. Proposals exist for JSON-RPC over HTTP in different versions, but they are not part of the spec.
This change implements an interface for the status code generating code and a default implementation reflecting the existing behaviour.
